refactor: consolidate broadcaster modules
