+++ crates/apps/src/bin/dsolver-tycho-broadcaster-service.rs
@@ -3,7 +3,7 @@ static GLOBAL: jemallocator::Jemalloc = jemallocator::Jemalloc;
 
 #[tokio::main]
 async fn main() -> anyhow::Result<()> {
-    let service = runtime::broadcaster_service::build_broadcaster_service().await?;
+    let service = runtime::broadcaster::app::build_broadcaster_service().await?;
     let addr = std::net::SocketAddr::from((service.config.host, service.config.port));
     let app = rpc::create_broadcaster_router(service.app_state);
