+++ crates/runtime/src/broadcaster/app.rs
@@ -8,6 +8,7 @@ use tycho_simulation::utils::load_all_tokens;
 use crate::broadcaster::redis_publisher::{
     BroadcasterRedisPublisher, BroadcasterRedisPublisherConfig, TokioRedisStreamWriter,
 };
+use crate::broadcaster::service::{BroadcasterServiceState, SnapshotSessionError};
 use crate::broadcaster::state::{
     BroadcasterReadiness, BroadcasterSnapshotCache, BroadcasterStatusSnapshot,
     BroadcasterUpstreamState,
@@ -22,9 +23,6 @@ use crate::models::tokens::{TokenStore, TokenStoreError};
 use crate::services::rfq_tokens::{load_rfq_token_stores, RfqTokenStoreConfig};
 use crate::services::stream_builder::{
     build_broadcaster_raw_stream, build_rfq_stream, BroadcasterProtocols, RFQConfig, RFQTokenStores,
-};
-use crate::services::{
-    broadcaster::BroadcasterServiceState, broadcaster_sessions::SnapshotSessionError,
 };
 use crate::stream::{
     supervise_broadcaster_raw_stream, supervise_broadcaster_stream, BroadcasterStreamControls,
