+++ crates/runtime/src/broadcaster/mod.rs
@@ -1,2 +1,5 @@
+pub mod app;
 pub mod redis_publisher;
+pub(crate) mod redis_subscription;
+pub mod service;
 pub mod state;
