+++ crates/runtime/src/simulator_service.rs
@@ -10,6 +10,11 @@ use tycho_simulation::tycho_common::{
     Bytes,
 };
 
+use crate::broadcaster::redis_subscription::{
+    supervise_broadcaster_redis_subscription, BroadcasterSubscriptionControls,
+    NativeBroadcasterSubscriptionControls, RfqBroadcasterSubscriptionControls,
+    VmBroadcasterSubscriptionControls,
+};
 use crate::config::{
     init_logging, load_broadcaster_redis_config, load_config, AppConfig, BroadcasterRedisConfig,
     MemoryConfig,
@@ -19,11 +24,6 @@ use crate::models::state::{AppState, BroadcasterSubscriptionStatus, StateStore, 
 use crate::models::stream_health::StreamHealth;
 use crate::models::tokens::{
     derive_broadcaster_token_lookup_url, derive_broadcaster_token_snapshot_url, TokenStore,
-};
-use crate::services::broadcaster_subscription::{
-    supervise_broadcaster_redis_subscription, BroadcasterSubscriptionControls,
-    NativeBroadcasterSubscriptionControls, RfqBroadcasterSubscriptionControls,
-    VmBroadcasterSubscriptionControls,
 };
 use crate::services::{EncodeService, QuoteService};
 use crate::stream::StreamSupervisorConfig;
